+++ SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/ModifyTableTranslation.java
@@ -128,7 +128,7 @@ class ModifyTableTranslation implements ICompilerComponent {
         if (this.columnPermutation == null)
             return source;
         DBSPZSetLiteral result = DBSPZSetLiteral.emptyWithElementType(this.resultType);
-        for (Map.Entry<DBSPExpression, Integer> e : source.data.entrySet()) {
+        for (Map.Entry<DBSPExpression, Long> e : source.data.entrySet()) {
             DBSPExpression perm = this.permuteColumns(e.getKey());
             result.add(perm, e.getValue());
         }
